third_party/WebKit/LayoutTests/media/track/opera/interfaces/VTTCue/vertical-expected.txt

 
PASS VTTCue.vertical, script-created cue 
FAIL VTTCue.vertical, parsed cue assert_equals: expected "rl" but got "lr"
